requirements.txt
streamlit==1.41.1
scikit-learn==1.5.2
python-docx==1.1.2
pypdf==5.1.0
reportlab==4.2.5
